refactor!: rename `TextDocumentSplitter` to `DocumentSplitter` (#6223)
* rename TextDocumentSplitter to DocumentSplitter

* reno

* fix init
