+++ haystack/preview/components/preprocessors/document_cleaner.py
