+++ releasenotes/notes/rename-text-document-splitter-1e9fcd292c4591dd.yaml
@@ -1,0 +1,6 @@
+---
+preview:
+  - |
+    rename `TextDocumentSplitter` to `DocumentSplitter`, to allow a better
+    distinction between Components that operate on text and those that operate
+    on Documents.
