+++ test/preview/components/preprocessors/test_document_splitter.py
@@ -1,48 +1,48 @@
 import pytest
 
 from haystack.preview import Document
-from haystack.preview.components.preprocessors import TextDocumentSplitter
+from haystack.preview.components.preprocessors import DocumentSplitter
 
 
-class TestTextDocumentSplitter:
+class TestDocumentSplitter:
     @pytest.mark.unit
     def test_non_text_document(self):
         with pytest.raises(
-            ValueError, match="TextDocumentSplitter only works with text documents but document.content for document ID"
+            ValueError, match="DocumentSplitter only works with text documents but document.content for document ID"
         ):
-            splitter = TextDocumentSplitter()
+            splitter = DocumentSplitter()
             splitter.run(documents=[Document()])
 
     @pytest.mark.unit
     def test_single_doc(self):
-        with pytest.raises(TypeError, match="TextDocumentSplitter expects a List of Documents as input."):
-            splitter = TextDocumentSplitter()
+        with pytest.raises(TypeError, match="DocumentSplitter expects a List of Documents as input."):
+            splitter = DocumentSplitter()
             splitter.run(documents=Document())
 
     @pytest.mark.unit
     def test_empty_list(self):
-        splitter = TextDocumentSplitter()
+        splitter = DocumentSplitter()
         res = splitter.run(documents=[])
         assert res == {"documents": []}
 
     @pytest.mark.unit
     def test_unsupported_split_by(self):
         with pytest.raises(ValueError, match="split_by must be one of 'word', 'sentence' or 'passage'."):
-            TextDocumentSplitter(split_by="unsupported")
+            DocumentSplitter(split_by="unsupported")
 
     @pytest.mark.unit
     def test_unsupported_split_length(self):
         with pytest.raises(ValueError, match="split_length must be greater than 0."):
-            TextDocumentSplitter(split_length=0)
+            DocumentSplitter(split_length=0)
 
     @pytest.mark.unit
     def test_unsupported_split_overlap(self):
         with pytest.raises(ValueError, match="split_overlap must be greater than or equal to 0."):
-            TextDocumentSplitter(split_overlap=-1)
+            DocumentSplitter(split_overlap=-1)
 
     @pytest.mark.unit
     def test_split_by_word(self):
-        splitter = TextDocumentSplitter(split_by="word", split_length=10)
+        splitter = DocumentSplitter(split_by="word", split_length=10)
         result = splitter.run(
             documents=[
                 Document(
@@ -56,7 +56,7 @@ class TestTextDocumentSplitter:
 
     @pytest.mark.unit
     def test_split_by_word_multiple_input_docs(self):
-        splitter = TextDocumentSplitter(split_by="word", split_length=10)
+        splitter = DocumentSplitter(split_by="word", split_length=10)
         result = splitter.run(
             documents=[
                 Document(
@@ -76,7 +76,7 @@ class TestTextDocumentSplitter:
 
     @pytest.mark.unit
     def test_split_by_sentence(self):
-        splitter = TextDocumentSplitter(split_by="sentence", split_length=1)
+        splitter = DocumentSplitter(split_by="sentence", split_length=1)
         result = splitter.run(
             documents=[
                 Document(
@@ -91,7 +91,7 @@ class TestTextDocumentSplitter:
 
     @pytest.mark.unit
     def test_split_by_passage(self):
-        splitter = TextDocumentSplitter(split_by="passage", split_length=1)
+        splitter = DocumentSplitter(split_by="passage", split_length=1)
         result = splitter.run(
             documents=[
                 Document(
@@ -106,7 +106,7 @@ class TestTextDocumentSplitter:
 
     @pytest.mark.unit
     def test_split_by_word_with_overlap(self):
-        splitter = TextDocumentSplitter(split_by="word", split_length=10, split_overlap=2)
+        splitter = DocumentSplitter(split_by="word", split_length=10, split_overlap=2)
         result = splitter.run(
             documents=[
                 Document(
@@ -120,7 +120,7 @@ class TestTextDocumentSplitter:
 
     @pytest.mark.unit
     def test_source_id_stored_in_metadata(self):
-        splitter = TextDocumentSplitter(split_by="word", split_length=10)
+        splitter = DocumentSplitter(split_by="word", split_length=10)
         doc1 = Document(content="This is a text with some words.")
         doc2 = Document(content="This is a different text with some words.")
         result = splitter.run(documents=[doc1, doc2])
@@ -129,7 +129,7 @@ class TestTextDocumentSplitter:
 
     @pytest.mark.unit
     def test_copy_metadata(self):
-        splitter = TextDocumentSplitter(split_by="word", split_length=10)
+        splitter = DocumentSplitter(split_by="word", split_length=10)
         documents = [
             Document(content="Text.", meta={"name": "doc 0"}),
             Document(content="Text.", meta={"name": "doc 1"}),
